+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Module `ProductSummaryTypes` to import types from other apps.
+
+### Fixed
+- Missing fields from types.
 
 ## [0.6.0] - 2020-11-05
 ### Added